feat: defkbd
Ports defkbd from clawe.bindings. This one is trickier - it supports
bindings to write out to both sxhkd and awesomewm, and we get away with
not depending on ralphie.awesome for `awm/awm-fnl` b/c it's quoted.
As with other defthing keys, there is some strong coupling between this
and code in russmatney/clawe.